+++ tests/test_output_formats.py
@@ -1,8 +1,11 @@
 import json
 from collections import OrderedDict
 
+from code42cli.output_formats import extraction_output_format
+from code42cli.output_formats import get_dynamic_header
 from code42cli.output_formats import output_format
 from code42cli.output_formats import to_csv
+from code42cli.output_formats import to_dynamic_csv
 from code42cli.output_formats import to_formatted_json
 from code42cli.output_formats import to_json
 from code42cli.output_formats import to_table
@@ -112,6 +115,31 @@ d12d54f0-5160-47a8-a48f-7d5fa5b051c5,outside td,HIGH,FED_CLOUD_SHARE_PERMISSIONS
 8b393324-c34c-44ac-9f79-4313601dd859,Test different filters,MEDIUM,FED_ENDPOINT_EXFILTRATION,Alerting,True
 5eabed1d-a406-4dfc-af81-f7485ee09b19,Test Alerts using CLI,HIGH,FED_ENDPOINT_EXFILTRATION,Alerting,True"""
 
+DYNAMIC_CSV_OUTPUT = "\r\n".join(
+    [
+        "observerRuleId,name,severity,type,ruleSource,isEnabled",
+        "d12d54f0-5160-47a8-a48f-7d5fa5b051c5,outside td,HIGH,FED_CLOUD_SHARE_PERMISSIONS,Alerting,True",
+        "8b393324-c34c-44ac-9f79-4313601dd859,Test different filters,MEDIUM,FED_ENDPOINT_EXFILTRATION,Alerting,True",
+        "5eabed1d-a406-4dfc-af81-f7485ee09b19,Test Alerts using CLI,HIGH,FED_ENDPOINT_EXFILTRATION,Alerting,True",
+        "",
+    ]
+)
+
+
+TEST_NESTED_DATA = {
+    "test": "TEST",
+    "name": "outside td",
+    "description": "",
+    "severity": "HIGH",
+    "isSystem": False,
+    "isEnabled": True,
+    "ruleSource": ["Alerting"],
+    "tenantId": "1d71796f-af5b-4231-9d8e-df6434da4663",
+    "observerRuleId": {"test": ["d12d54f0-5160-47a8-a48f-7d5fa5b051c5"]},
+    "type": ["FED_CLOUD_SHARE_PERMISSIONS"],
+    "id": "5157f1df-cb3e-4755-92a2-0f42c7841020",
+}
+
 
 def test_to_csv_formats_data_to_csv_format():
     formatted_output = to_csv(TEST_DATA, TEST_HEADER)
@@ -126,7 +154,7 @@ def test_to_table_formats_data_to_table_format():
 
 def test_to_json():
     formatted_output = to_json(TEST_DATA, TEST_HEADER)
-    assert formatted_output == json.dumps(FILTERED_OUTPUT)
+    assert formatted_output == json.dumps(TEST_DATA)
 
 
 def test_to_formatted_json():
@@ -157,3 +185,40 @@ def test_output_format_returns_to_csv_function_when_csv_format_option_is_passed(
 def test_output_format_returns_to_table_function_when_no_format_option_is_passed():
     format_function = output_format(None, None, None)
     assert id(format_function) == id(to_table)
+
+
+def test_output_format_returns_to_dynamic_csv_function_when_csv_option_is_passed():
+    extraction_output_format_function = extraction_output_format(None, None, "CSV")
+    assert id(extraction_output_format_function) == id(to_dynamic_csv)
+
+
+def test_output_format_returns_to_table_function_when_table_option_is_passed():
+    extraction_output_format_function = extraction_output_format(None, None, "TABLE")
+    assert id(extraction_output_format_function) == id(to_table)
+
+
+def test_extraction_output_format_returns_to_formatted_json_function_when_json__option_is_passed():
+    format_function = extraction_output_format(None, None, "JSON")
+    assert id(format_function) == id(to_formatted_json)
+
+
+def test_extraction_output_format_returns_to_json_function_when_raw_json_format_option_is_passed():
+    format_function = extraction_output_format(None, None, "RAW-JSON")
+    assert id(format_function) == id(to_json)
+
+
+def test_get_format_header_returns_all_keys_only_which_are_not_nested():
+    header = get_dynamic_header(TEST_NESTED_DATA)
+    assert header == {
+        "test": "Test",
+        "name": "Name",
+        "description": "Description",
+        "severity": "Severity",
+        "tenantId": "Tenantid",
+        "id": "Id",
+    }
+
+
+def test_to_dynamic_csv():
+    formatted_output = to_dynamic_csv(TEST_DATA, TEST_HEADER)
+    assert formatted_output == DYNAMIC_CSV_OUTPUT